pyproject.toml
[project]
name = "telegram-forwarder"
version = "1.0.0"
description = "A Telegram bot that forwards messages from Contact A to Contact B"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyrogram>=2.0.106",
    "TgCrypto>=1.2.5",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
